workflow: add test suite
