+++ package.json
@@ -78,7 +78,7 @@
     "semver": "^7.3.5",
     "ts-jest": "27.0.7",
     "tslib": "2.3.1",
-    "tsup": "5.12.8",
+    "tsup": "5.7.2",
     "typescript": "4.6.4",
     "wait-on": "6.0.0",
     "wireit": "^0.4.3",
