+++ packages/test-suite/tsup.config.ts
@@ -1,0 +1,4 @@
+import pkg from './package.json';
+import { baseTsup } from '../../tsup.config';
+
+export const tsup = baseTsup(pkg);
