+++ packages/utils/src/index.ts
@@ -1,4 +1,3 @@
-export * from './test';
 export * from './utils';
 export * from './queue';
 export * from './sentry';
